Consolidate rpc_ctx definitions in rpc_ctx files
Initialize empty rpc_msg in rpc_ctx tree (bugfix).

Move rpc_msg_init() into rpc_ctx.c (where it is now used) instead of svc.c
(where it was not used).  Likewise, move its header into rpc_ctx.h.

Move rpc_ctx_t to rpc_ctx.h (where it is used).

Remove 4 unused obsolete defined flags.
