+++ src/svc_internal.h
@@ -204,7 +204,6 @@ epoll_create_wr(size_t size, int flags)
 extern void __rpc_set_blkin_endpoint(SVCXPRT *xprt, const char *tag);
 #endif
 
-void rpc_msg_init(struct rpc_msg *msg);
 void svc_rqst_shutdown(void);
 
 #endif				/* TIRPC_SVC_INTERNAL_H */
